+++ NEWS.md
@@ -1,3 +1,9 @@
+# fwebinfr devel 
+
+* `fw_infer()` gains arguments `sdB` to dealth with biomass variation and `...`
+to pass futher argument to `limSolve::xsample()`.
+
+
 # fwebinfr 0.0.1
 
 * `fw_infer()` performs LIM for a system described by a generalized Lotka-Voltera equation set.